Mock Drivers Refactor
This path refactors the mock (and bad mock) drivers formerly in the
`github.com/emccode/rexray/test` package into the
`github.com/emccode/rexray/drivers/mock` package. This is to support
mock drivers during normal operations and not just automated tests.

The mock drivers are not automatically registered. To elect to use the
mock drivers an environment variable of `REXRAY_MOCKDRIVERS` must be
set to a value that evaluates to boolean true.

At this point it is possible to use `mockOSDriver`, `mockVolumeDriver`,
and `mockStorageDriver` as values for the environment variables, config
options, and CLI flags that control which drivers should be initialized.
