+++ drivers/drivers.go
@@ -2,6 +2,7 @@ package drivers
 
 import (
 	// loads the drivers
+	_ "github.com/emccode/rexray/drivers/mock"
 	_ "github.com/emccode/rexray/drivers/os"
 	_ "github.com/emccode/rexray/drivers/storage"
 	_ "github.com/emccode/rexray/drivers/volume"
