+++ test/bad_mock_drivers_test.go
@@ -5,13 +5,14 @@ import (
 
 	"github.com/emccode/rexray/core"
 	"github.com/emccode/rexray/core/errors"
+	"github.com/emccode/rexray/drivers/mock"
 )
 
 func TestNewWithBadOSDriver(t *testing.T) {
 	r := core.New(nil)
-	r.Config.OSDrivers = []string{badMockOSDriverName}
-	r.Config.VolumeDrivers = []string{mockVolDriverName}
-	r.Config.StorageDrivers = []string{mockStorDriverName}
+	r.Config.OSDrivers = []string{mock.BadMockOSDriverName}
+	r.Config.VolumeDrivers = []string{mock.MockVolDriverName}
+	r.Config.StorageDrivers = []string{mock.MockStorDriverName}
 	if err := r.InitDrivers(); err != errors.ErrNoOSDrivers {
 		t.Fatal(err)
 	}
@@ -19,9 +20,9 @@ func TestNewWithBadOSDriver(t *testing.T) {
 
 func TestNewWithBadVolumeDriver(t *testing.T) {
 	r := core.New(nil)
-	r.Config.OSDrivers = []string{mockOSDriverName}
-	r.Config.VolumeDrivers = []string{badMockVolDriverName}
-	r.Config.StorageDrivers = []string{mockStorDriverName}
+	r.Config.OSDrivers = []string{mock.MockOSDriverName}
+	r.Config.VolumeDrivers = []string{mock.BadMockVolDriverName}
+	r.Config.StorageDrivers = []string{mock.MockStorDriverName}
 	if err := r.InitDrivers(); err != errors.ErrNoVolumeDrivers {
 		t.Fatal(err)
 	}
@@ -29,52 +30,10 @@ func TestNewWithBadVolumeDriver(t *testing.T) {
 
 func TestNewWithBadStorageDriver(t *testing.T) {
 	r := core.New(nil)
-	r.Config.OSDrivers = []string{mockOSDriverName}
-	r.Config.VolumeDrivers = []string{mockVolDriverName}
-	r.Config.StorageDrivers = []string{badMockStorDriverName}
+	r.Config.OSDrivers = []string{mock.MockOSDriverName}
+	r.Config.VolumeDrivers = []string{mock.MockVolDriverName}
+	r.Config.StorageDrivers = []string{mock.BadMockStorDriverName}
 	if err := r.InitDrivers(); err != errors.ErrNoStorageDrivers {
 		t.Fatal(err)
 	}
-}
-
-type badMockOSDriver struct {
-	mockOSDriver
-}
-
-func newBadOSDriver() core.Driver {
-	var d core.OSDriver = &badMockOSDriver{
-		mockOSDriver{badMockOSDriverName}}
-	return d
-}
-
-func (m *badMockOSDriver) Init(r *core.RexRay) error {
-	return errors.New("init error")
-}
-
-type badMockVolDriver struct {
-	mockVolDriver
-}
-
-func newBadVolDriver() core.Driver {
-	var d core.VolumeDriver = &badMockVolDriver{
-		mockVolDriver{badMockVolDriverName}}
-	return d
-}
-
-func (m *badMockVolDriver) Init(r *core.RexRay) error {
-	return errors.New("init error")
-}
-
-type badMockStorDriver struct {
-	mockStorDriver
-}
-
-func newBadStorDriver() core.Driver {
-	var d core.StorageDriver = &badMockStorDriver{
-		mockStorDriver{badMockStorDriverName}}
-	return d
-}
-
-func (m *badMockStorDriver) Init(r *core.RexRay) error {
-	return errors.New("init error")
 }
